User supplied room ids
Move to v6 api

+++ lib/chatkit/client.rb
@@ -42,7 +42,7 @@ module Chatkit
       @api_instance = PusherPlatform::Instance.new(
         base_options.merge({
           service_name: 'chatkit',
-          service_version: 'v4'
+          service_version: 'v6'
         })
       )
 
@@ -238,6 +238,7 @@ module Chatkit
         private: options[:private] || false
       }
 
+      body[:id] = options[:id] unless options[:id].nil?
       body[:custom_data] = options[:custom_data] unless options[:custom_data].nil?
 
       unless options[:user_ids].nil?
@@ -503,9 +504,13 @@ module Chatkit
         raise Chatkit::MissingParameterError.new("You must provide the ID of the message you want to delete")
       end
 
+      if options[:room_id].nil?
+        raise Chatkit::MissingParameterError.new("You must provide the ID of the room to which the message belongs")
+      end
+
       api_request({
         method: "DELETE",
-        path: "/messages/#{options[:id]}",
+        path: "/messages/rooms/#{options[:room_id]}/#{options[:id]}",
         jwt: generate_su_token[:token]
       })
     end
